test: 공통 메서드 분리
